Changed headers from Integer to long -- patient IDs are typed as Long

+++ UnifiedPatientManager/src/main/java/com/Eges411Team/UnifiedPatientManager/repositories/AllergyRepository.java
@@ -4,7 +4,7 @@ import com.Eges411Team.UnifiedPatientManager.entity.Allergy;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface AllergyRepository extends JpaRepository<Allergy, Integer> {
+public interface AllergyRepository extends JpaRepository<Allergy, Long> {
 
     // Uses the Java field name "patient_id"
     List<Allergy> findAllByPatient_id(Long patient_id);
